+++ tests/examples/astrog_test.py
@@ -21,8 +21,6 @@ import matplotlib.pyplot as plt
 plt.close('all')
 import hatyan
 
-hatyan.init_RWS()
-
 #dir_testdata = 'P:\\1209447-kpp-hydraulicaprogrammatuur\\hatyan\\hatyan_data_acceptancetests'
 dir_testdata = 'C:\\DATA\\hatyan_data_acceptancetests'
 
@@ -134,5 +132,3 @@ if compare2fortran:
     fig, (ax1,ax2,ax3) = hatyan.plot_astrog_diff(pd_python=seasons_python, pd_fortran=seasons_fortran, typeLab=['spring','summer','autumn','winter'], timeBand=[-30,30])
     fig.savefig('season_differences.png')
 
-#%%
-hatyan.exit_RWS() #provides footer to outputfile when calling this script with python