+++ tests/examples/predictie_2019_allfromdia.py
@@ -15,7 +15,6 @@ import datetime as dt
 import pandas as pd
 import hatyan
 
-hatyan.init_RWS()
 #dir_testdata = 'P:\\1209447-kpp-hydraulicaprogrammatuur\\hatyan\\hatyan_data_acceptancetests'
 dir_testdata = 'C:\\DATA\\hatyan_data_acceptancetests'
 
@@ -130,11 +129,7 @@ for current_station in selected_stations:
     #fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_measurements_group0)
     #fig.savefig('prediction_%im_%s_measurements'%(times_step_pred, current_station))
 
-        
-hatyan.exit_RWS()
 print('\nthese %i stations were requested for processing:\n%s'%(len(selected_stations),selected_stations))
 print('\nthese %i stations were not processed because there is no ana/comp dataset available:\n%s'%(len(stats_noana),stats_noana))
 print('\nthese %i stations were not processed because there is no 4Y or likewise data available:\n%s'%(len(stats_no4Y),stats_no4Y))
 print('\nthese %i stations were not processed because there is no 19Y data available:\n%s'%(len(stats_no19Y),stats_no19Y))
-
-