+++ tests/examples/predictie_2019_frommergedcomp.py
@@ -12,7 +12,6 @@ import os
 import datetime as dt
 import hatyan
 
-hatyan.init_RWS()
 #dir_testdata = 'P:\\1209447-kpp-hydraulicaprogrammatuur\\hatyan\\hatyan_data_acceptancetests'
 dir_testdata = 'C:\\DATA\\hatyan_data_acceptancetests'
 
@@ -86,10 +85,5 @@ for current_station in selected_stations:
     fig, (ax1,ax2) = hatyan.plot_timeseries(ts=ts_prediction, ts_validation=ts_validation)
     fig.savefig('prediction_%im_%s_validation'%(times_step_pred, current_station))
 
-    
-hatyan.exit_RWS()
 print('\nthese %i stations were requested for processing:\n%s'%(len(selected_stations),selected_stations))
 print('\nthese %i stations were not processed because there is no ana/comp dataset available:\n%s'%(len(stats_noana),stats_noana))
-    
-    
-